requirements.txt
annotated-types==0.6.0 ; python_version >= "3.11" and python_version < "3.13"
anyio==3.7.1 ; python_version >= "3.11" and python_version < "3.13"
async-timeout==4.0.3 ; python_version >= "3.11" and python_full_version <= "3.11.2"
black==23.11.0 ; python_version >= "3.11" and python_version < "3.13"
certifi==2023.11.17 ; python_version >= "3.11" and python_version < "3.13"
cffi==1.16.0 ; python_version >= "3.11" and python_version < "3.13"
click==8.1.7 ; python_version >= "3.11" and python_version < "3.13"
colorama==0.4.6 ; python_version >= "3.11" and python_version < "3.13" and (sys_platform == "win32" or platform_system == "Windows")
coverage==7.3.2 ; python_version >= "3.11" and python_version < "3.13"
<<<<<<< HEAD
=======
coverage[toml]==7.3.2 ; python_version >= "3.11" and python_version < "3.13"
>>>>>>> origin/main
cryptography==41.0.7 ; python_version >= "3.11" and python_version < "3.13"
fastapi==0.104.1 ; python_version >= "3.11" and python_version < "3.13"
h11==0.14.0 ; python_version >= "3.11" and python_version < "3.13"
httpcore==1.0.2 ; python_version >= "3.11" and python_version < "3.13"
httpx==0.25.2 ; python_version >= "3.11" and python_version < "3.13"
idna==3.6 ; python_version >= "3.11" and python_version < "3.13"
iniconfig==2.0.0 ; python_version >= "3.11" and python_version < "3.13"
libnum==1.7.1 ; python_version >= "3.11" and python_version < "3.13"
mypy-extensions==1.0.0 ; python_version >= "3.11" and python_version < "3.13"
numpy==1.26.2 ; python_version >= "3.11" and python_version < "3.13"
packaging==23.2 ; python_version >= "3.11" and python_version < "3.13"
pathspec==0.11.2 ; python_version >= "3.11" and python_version < "3.13"
platformdirs==4.1.0 ; python_version >= "3.11" and python_version < "3.13"
pluggy==1.3.0 ; python_version >= "3.11" and python_version < "3.13"
pokerlib==2.2.7 ; python_version >= "3.11" and python_version < "3.13"
pycparser==2.21 ; python_version >= "3.11" and python_version < "3.13"
pydantic-core==2.14.5 ; python_version >= "3.11" and python_version < "3.13"
pydantic==2.5.2 ; python_version >= "3.11" and python_version < "3.13"
pytest-asyncio==0.21.1 ; python_version >= "3.11" and python_version < "3.13"
pytest-cov==4.1.0 ; python_version >= "3.11" and python_version < "3.13"
pytest==7.4.3 ; python_version >= "3.11" and python_version < "3.13"
<<<<<<< HEAD
=======
python-dotenv==1.0.0 ; python_version >= "3.11" and python_version < "3.13"
>>>>>>> origin/main
redis==5.0.1 ; python_version >= "3.11" and python_version < "3.13"
sniffio==1.3.0 ; python_version >= "3.11" and python_version < "3.13"
starlette==0.27.0 ; python_version >= "3.11" and python_version < "3.13"
treys==0.1.8 ; python_version >= "3.11" and python_version < "3.13"
typing-extensions==4.8.0 ; python_version >= "3.11" and python_version < "3.13"
uvicorn==0.24.0.post1 ; python_version >= "3.11" and python_version < "3.13"
websockets==12.0 ; python_version >= "3.11" and python_version < "3.13"
